+++ R/input-validation.R
@@ -74,8 +74,14 @@ input_validation_anticlustering <- function(x, K, objective, method,
   }
 
   if (!inherits(objective, "function")) {
-    validate_input(objective, "objective", input_set = c("distance", "diversity", "variance", "kplus"), 
-                   len = 1, not_na = TRUE)
+    validate_input(objective, "objective", input_set = c(
+      "distance", 
+      "diversity", 
+      "dispersion",
+      "variance", 
+      "kplus"
+      ), 
+      len = 1, not_na = TRUE)
     if (objective == "variance" && method == "ilp") {
       stop("You cannot use integer linear programming method to maximize the variance criterion. ",
            "Use objective = 'distance', or method = 'exchange' instead")